Replace javascript alert with sweetalert2

+++ frontend/src/admin/views/SubAdmin.js
@@ -54,26 +54,29 @@ const SubAdmin = () => {
                 action.resetForm();
                 getSubAdmins();
                 handleClose();
-                alert(data.message);
+                config.alerts.success(data.message);
             } else {
-                alert(data.message);
+                config.alerts.error(data.message);
             }
         }
     });
 
     const removeSubAdmin = async (id) => {
-        const response = await fetch(`${config.SERVER_URL}/api/remove-sub-admin/${id}`, {
-            method: "DELETE",
-            headers: {
-                "Authorization": `bearer ${Cookies.get("jwt")}`
+        const isConfirmed = await config.alerts.confirm("Are you sure?", "This action cannot be undone.");
+        if (isConfirmed) {
+            const response = await fetch(`${config.SERVER_URL}/api/remove-sub-admin/${id}`, {
+                method: "DELETE",
+                headers: {
+                    "Authorization": `bearer ${Cookies.get("jwt")}`
+                }
+            });
+            const data = await response.json();
+            if (response.ok) {
+                getSubAdmins();
+                config.alerts.success(data.message);
+            } else {
+                config.alerts.error(data.message);
             }
-        });
-        const data = await response.json();
-        if (response.ok) {
-            getSubAdmins();
-            alert(data.message);
-        } else {
-            alert(data.message);
         }
     }
 
@@ -87,10 +90,10 @@ const SubAdmin = () => {
                 setSubAdmin(data.data);
                 setTotalSubAdmin(data.totalSubAdmins);
             } else {
-                alert(data.message);
+                config.alerts.error(data.message);
             }
         } catch (error) {
-            alert(error.toString());
+            config.alerts.error(error.toString());
             console.log(error);
         }
     }, [activePage]);
